Implement accounts index cachedb as optional flag '--accounts-index'
This enables most users to use the standard memory-caching way of handling a small-medium number of accounts.
Power users can activate flag to create and maintain idempotent cachedb for indexing the accounts.

TODO:
There are still a few bugs in syncfs2db, where cachedb should sync efficiently with the given FS.
For example, removed files are not removed from the cachedb consistently."