solution: make rotateOld thread safe (and KISS)
problem: when rotation of old files take a long time (compression is
turned on, there are lots of large old files to rotate, etc) and there
is new rotation request (current log file size or time based) it is
possible, that rotateOld will be executed twice in parallel, and list
of files gathered in one execution can be outdated by changes made by
another execution.

comment: log handling is very conservative to avoid any loss of logs.
If there is a problem with opening a file (for example for compression),
fatal error is reported.
Following KISS principle, mutex is added to protect the entire rotateOld.