sync: eth/downloader an incomplete sketch of downloader
signature modification and implementation of BLockchaind
and Lightchain interfaces.

I've tried to isolate these changes, but they are deeply involved
with the next commit's changes and separating the concerns is beyond
my patience right now.

Be advised: the downloader fetcher and fast sync process will be modified
to make use of this revision's structural change.